+++ packages/xchain-litecoin/__tests__/client.test.ts
@@ -50,6 +50,14 @@ describe('LitecoinClient Test', () => {
     expect(ltcClient.setPhrase(phraseOne)).toBeUndefined
   })
 
+  it('should not throw on a client without a phrase', () => {
+    expect(() => {
+      new Client({
+        network: Network.Testnet,
+      })
+    }).not.toThrow()
+  })
+
   it('should validate the right address', () => {
     ltcClient.setNetwork(Network.Testnet)
     ltcClient.setPhrase(phraseOne)
@@ -307,5 +315,16 @@ describe('LitecoinClient Test', () => {
     expect(ltcClient.getExplorerTxUrl('anotherTestTxHere')).toEqual(
       'https://blockexplorer.one/litecoin/testnet/blockHash/anotherTestTxHere',
     )
+  })
+
+  it('should broadcast a deposit to thorchain inbound address', async () => {
+    ltcClient.setNetwork(Network.Testnet)
+    ltcClient.setPhrase(phraseOne)
+    const txid = await ltcClient.deposit({
+      asset: AssetLTC,
+      amount: baseAmount(2223),
+      memo: '=:THOR.RUNE:tthor1puhn8fclwvmmzh7uj7546wnxz5h3zar8e66sc5',
+    })
+    expect(txid).toEqual('mock-txid')
   })
 })