+++ packages/xchain-bitcoincash/CHANGELOG.md
@@ -1,3 +1,12 @@
+# v.0.13.1 (2022-05-05)
+
+## Update
+
+- Add `deposit` function to BitcoinCash `Client`
+- Update latest dependencies
+- Add tests for `deposit` 
+
+
 # v.0.13.0 (2022-03-23)
 
 ## Add